Improve testing coverage, remove unused code
a) Add further tests to CI to cover most features. This uncovered a potential race
in case setoption was sent between two searches. As the UCI protocol requires
this sent to be went the engine is not searching, setoption now ensures that
this is the case.

b) Remove some unused code

closes #4730

No functional change
 master
 stockfish-dev-20230811-81929458
@vondele
@snicolet
vondele authored